Price handled as int but shown as float.

See merge request samulu/handle-that-pos!37

+++ src/main/java/model/interfaces/IOrder.java
@@ -13,5 +13,5 @@ public interface IOrder {
     boolean addProductToOrder(Product product);
     boolean removeProductFromOrder(Product product);
     ArrayList<Product> getProductList();
-    float getTotalPrice();
+    int getTotalPrice();
 }
